Remove usage of old check for drawer
Previously, we made use of Scaffold.of(context).hasDrawer. This works, but the screen is not updated when resized, for example. It was migrated to Scaffold.hasDrawer, which updates the underlaying widgets accordingly.

Also, removed other scaffold initializations in the app, which was causing the app to not open the drawer properly.

Created UnityDrawerButton, a M3 drawer button that is compact to eyes and work for every platform.