feat: safe loading unity providers
Sometimes, safe_local_storage attempts to use a version of a provider - but it may be, for some reason, not available. If so, it throws an error. SInce the providers are loaded before the app starts, it may stop the app from starting.

WIth the add of UnityProvider, the error is properly handled and should not crash the app.